refactor: remove showError to class ErrorView
