+++ async-race/src/app/view/pages/garage/Garage.ts
@@ -2,6 +2,7 @@ import API from '../../../../api/Api';
 import { TColorHEX } from '../../../../typescript/types';
 import PageBuilder from '../../../utils/PageBuilder';
 import Editor from './editor/Editor';
+
 import GarageTable from './table/GarageTable';
 
 export default class Garage {
